Report on Design and Fabrication of Novel Multipurpose Robot
I'm currently working on this project which is based on robotics and automation. This project comprises of a novel multi purpose semi automated robot which works on the principle of IR sensor detection. This robot will be able to follow a line, it can avoid the obstacles which it will encounter in its path and can will also be able to follow the obstacle if required. This robot is basically a numerically controlled type of robot which works upon the set of coded instructions given to it by the programmer.