Add files via upload
In the era of the overcrowded (data communication) world, Li-Fi is a new way of wireless communication that uses LED lights to transmit data wirelessly. Transmission of data is one of the most important day-to-day activities in the fast-growing world. The current wireless networks that connect us to the Internet are very slow when multiple devices are connected. Also with the increase in the number of devices that access the Internet, the availability of fixed bandwidth makes it much more difficult to enjoy high data transfer rates and to connect a secure network. Radio waves are just a small part of the electromagnetic spectrum available for data transfer. Li-Fi has got a much broader spectrum for transmission compared to conventional methods of wireless communications that rely on radio waves. The basic ideology behind this technology is that the data can be transferred through LED light by varying light intensities faster than the human eyes can perceive. This technology uses a part of the electromagnetic spectrum that is still not greatly utilized- The Visible Spectrum, instead of Gigahertz radio waves for data transfer.