Design and Fabrication of Sensor Guided Robot
I'm currently working on this project which is based on robotics and automation. This project comprises a novel multi-purpose semi-automated robot that works on sensor detection. This robot will be able to follow a line, it can avoid the obstacles which it will encounter in its path and can also follow the obstacle when detected. This project aims to implement the algorithm and control the movement of the robot. It can be used as Industrial Automated Equipment Carrier, Small Household Application, and Guidance Application.